requirements.txt
discord.py
colorama
numpy
requests
aiohttp
bs4
pymongo
selenium
pyPrivnote
gtts
dnspython
image
termcolor
discord_webhooks
pysocks
licensing
tqdm
cfscrape
wget
scapy
dhooks
auto-py-to-exe
